Update v2.2.x for v2.2.3 release

+++ src/mca/ptl/tcp/ptl_tcp_component.c
@@ -244,6 +244,7 @@ static int component_register(void)
 }
 
 static char *urifile = NULL;
+static bool created_rendezvous_file = false;
 
 static pmix_status_t component_open(void)
 {
@@ -301,7 +302,9 @@ pmix_status_t component_close(void)
         free(mca_ptl_tcp_component.pid_filename);
     }
     if (NULL != mca_ptl_tcp_component.rendezvous_filename) {
-        unlink(mca_ptl_tcp_component.rendezvous_filename);
+        if (created_rendezvous_file) {
+            unlink(mca_ptl_tcp_component.rendezvous_filename);
+        }
         free(mca_ptl_tcp_component.rendezvous_filename);
     }
     if (NULL != urifile) {
@@ -357,8 +360,8 @@ static pmix_status_t setup_listener(pmix_info_t info[], size_t ninfo,
     bool session_tool = false;
     bool system_tool = false;
     pmix_socklen_t addrlen;
-    char *prefix, myhost[PMIX_MAXHOSTNAMELEN];
-    char myconnhost[PMIX_MAXHOSTNAMELEN];
+    char *prefix, myhost[PMIX_MAXHOSTNAMELEN] = {0};
+    char myconnhost[PMIX_MAXHOSTNAMELEN] = {0};
     int myport;
     pmix_kval_t *urikv;
 
@@ -618,17 +621,17 @@ static pmix_status_t setup_listener(pmix_info_t info[], size_t ninfo,
         goto sockerror;
     }
 
-    gethostname(myhost, sizeof(myhost));
+    gethostname(myhost, sizeof(myhost)-1);
     if (AF_INET == mca_ptl_tcp_component.connection.ss_family) {
         prefix = "tcp4://";
         myport = ntohs(((struct sockaddr_in*) &mca_ptl_tcp_component.connection)->sin_port);
         inet_ntop(AF_INET, &((struct sockaddr_in*) &mca_ptl_tcp_component.connection)->sin_addr,
-                  myconnhost, PMIX_MAXHOSTNAMELEN);
+                  myconnhost, PMIX_MAXHOSTNAMELEN-1);
     } else if (AF_INET6 == mca_ptl_tcp_component.connection.ss_family) {
         prefix = "tcp6://";
         myport = ntohs(((struct sockaddr_in6*) &mca_ptl_tcp_component.connection)->sin6_port);
         inet_ntop(AF_INET6, &((struct sockaddr_in6*) &mca_ptl_tcp_component.connection)->sin6_addr,
-                  myconnhost, PMIX_MAXHOSTNAMELEN);
+                  myconnhost, PMIX_MAXHOSTNAMELEN-1);
     } else {
         goto sockerror;
     }
@@ -708,6 +711,7 @@ static pmix_status_t setup_listener(pmix_info_t info[], size_t ninfo,
             mca_ptl_tcp_component.rendezvous_filename = NULL;
             goto sockerror;
         }
+        created_rendezvous_file = true;
     }
 
     /* if we are going to support tools, then drop contact file(s) */
@@ -873,7 +877,7 @@ static char **split_and_resolve(char **orig_str, char *name)
 {
     int i, ret, save, if_index;
     char **argv, *str, *tmp;
-    char if_name[IF_NAMESIZE];
+    char if_name[PMIX_IF_NAMESIZE];
     struct sockaddr_storage argv_inaddr, if_inaddr;
     uint32_t argv_prefix;
